Converting docs to Github markdown
The code in the docs doesn't highlight in Github, and on inspection it's because the code is wrapped in `pre` tags instead of using Github's code/syntax highlighting. I updated all the `pre` blocks and converted most URLs to use markdown style formatting (with the exception of the image wrapped in a link, since that still requires raw HTML).


Former-commit-id: 9c366420ac8594def6d024869ed0f0cddd9eb36d [formerly bf08fb07160362c0ebdd918b953ad3d1292ce70d]
Former-commit-id: a9fbb744a875988045e428e968b9ef6d63359347